refactor: move <SingleSelect> to /ui
